+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# 1.2.4
+- Fixes #37
+
+# 1.2.2
+- Added documentation keyword to pubspec.yaml
+
+# 1.2.1
+- Tweaks to improve our pub.dev score.
+
+# 1.2.0
+- removed the need to pass a version to the onepub add command.
+
+# 1.1.12
+- import command now sets the organisation name.
+- change logout command so it succeeds even if you have already logged out.
+- change import to use the new end point /organisation/token.
+- fixes for windows.
+
 # 1.1.11
 - Added a --ask switch to import so a user could directly enter the onepub token. This was to make it easy for ssh users to authenticate.
 - improvements to the output of onepub doctor.